# lusevokteren-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FjordVind Lusevokteren API
  description: |
    REST API for sea lice monitoring in Norwegian aquaculture.

    ## Authentication
    Most endpoints require authentication via JWT Bearer token.

    ```
    Authorization: Bearer <your-token>
    ```

    ## Demo Mode
    In development, demo tokens are available:
    - `demo_token_admin` - Admin access
    - `demo_token_driftsleder` - Operations manager
    - `demo_token_røkter` - Fish farmer

  version: 1.1.0
  contact:
    name: FjordVind Support
    email: support@fjordvind.no
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.fjordvind.no
    description: Production server
...
paths:
...
  /api/auth/register:
    post:
      summary: Register new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '409':
          description: User already exists
  /api/auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /api/auth/me:
    get:
      summary: Get current user
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
  /api/auth/refresh:
    post:
      summary: Refresh token
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: New token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  message:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - full_name
      properties:
        email:
          type: string
          format: email
          example: user@example.no
        password:
          type: string
          minLength: 8
          example: securepassword123
        full_name:
          type: string
          example: Ola Nordmann
        company_name:
          type: string
          example: Fjordbruk AS
        org_number:
          type: string
          pattern: '^\d{9}$'
          example: '123456789'
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [admin, driftsleder, røkter, viewer]
        company_id:
          type: string
          format: uuid
    UserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Valideringsfeil
        details:
          type: array
          items:
            type: string
...